feat: add more tests
