+++ src/__tests__/setup.ts
@@ -1,13 +1,14 @@
-import { afterAll, afterEach, beforeAll } from "vitest";
+import { afterAll, afterEach, beforeAll, vi } from "vitest";
 import { cleanup } from "@testing-library/react";
 
 import "@testing-library/jest-dom/vitest";
-// import "@testing-library/jest-dom";
-// import "@testing-library/jest-dom/extend-expect";
 
 import { server } from "./server";
+import LocalStorageMock from "../__mocks__/localStorage";
 
-beforeAll(() => server.listen());
+beforeAll(() => {
+  server.listen();
+});
 
 // runs a cleanup after each test case (e.g. clearing jsdom)
 afterEach(() => {
@@ -16,27 +17,6 @@ afterEach(() => {
 });
 
 afterAll(() => server.close());
-
-// Mock local storage
-class LocalStorageMock {
-  private store: Record<string, string> = {};
-
-  getItem(key: string): string | null {
-    return this.store[key] || null;
-  }
-
-  setItem(key: string, value: string): void {
-    this.store[key] = value.toString();
-  }
-
-  removeItem(key: string): void {
-    delete this.store[key];
-  }
-
-  clear(): void {
-    this.store = {};
-  }
-}
 
 Object.defineProperty(window, "localStorage", {
   value: new LocalStorageMock(),
