+++ src/pages/404.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 function NotFound() {
   return (
     <div
